requirements.txt
# Minimal requirements for the starter RAG script
sentence-transformers>=2.2.0        # embedding model utility (pulls in PyTorch)
faiss-cpu>=1.7.0                    # vector index for nearest-neighbor search
llama-cpp-python>=0.1.0             # python bindings for llama.cpp (local LLM inference)
numpy>=1.23.0
huggingface_hub
tqdm
